+++ tests/NeuralNet/Initializers/Xavier2NormalTest.php
@@ -140,7 +140,7 @@ final class Xavier2NormalTest extends TestCase
         $this->assertThat(
             $std,
             $this->logicalAnd(
-                $this->greaterThan($expectedStd * 0.83),
+                $this->greaterThan($expectedStd * 0.80),
                 $this->lessThan($expectedStd * 1.3)
             ),
             'Standard deviation does not match Xavier-2 Normal initialization'