+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     - Remove loss-based window early stopping
     - Add disk-based neural network snapshotting
     - Update Amp Backend to Amp version 2.0
+    - Add per-class smoothing to Gaussian Naive Bayes
     
 - 2.5.9
     - Symmetrize TSNE affinities matrix, fixes optimization